Modified some stuff related to communication
It seems like previously the JSON messages where 'malformed' (they where valid JSON objects but their structure was weird). Part of JSON message was structured properly (and could be read like a JSON object), but the other part was simply a string (and needed to be additionally parsed). Now the whole message is serialized and treated like a JSON object.