Reorganized some stuff, created a template git hook and updated the README
Created an empty pre-commit hook (originally intended to use it for a checklist that would show when trying to commit to 'main' branch, but changed my mind). Moved it to a separate folder, also moved AHK script to a separate folder. Later found out that the hook is not working when committing from GitHub Desktop, but I guess I'll still leave it there. Updated the README with info about the current repo structure.