Working on a big rewrite
I am not very happy with how the app is structured, so I'm rewriting a lot of it's networking code. The biggest change is that I'm moving almost all logic that manages app's state to the Flutter frontend. This removes a lot of duplicate code (the Soundboard class for example) and makes the bloc pattern more tight.
Before, every time something in UI needed to be changed, the UI would send a command to host to analyze the change. The host would eventually send a command back telling the UI what exactly needs to change. This 'split-in-half-change-handling' makes the code more messy and loose, which I don't like. Now, the UI is almost fully independent and manages it's state however it wants. All requests to the host are handled immediately and the results can be processed in the same place. Meanwhile all important events like errors are delivered asynchronously whenever they happen.
The rewrite is not finished yet, I just want to save my progress for the day.