Yes
I think the networking code is finally good. App seems to be stable and all parameters are handled properly. Also fixed a typo in English localization and the Playback error category. Made the Audio Engine Start/Stop button grayed out when the Audio Device and Sample Rate are not set