Reworked how the whole client-server thing operates
Previously, the UI was essentially a webpage, hosted inside the WPF window. Now UI and host are separate processes. Instead of showing a browser window, host now shows a window with logs that automatically hides when UI is launched. Other changes: added a simple workflow (that probably doesn't work, but we'll see), made a few changes to the UI and networking logic.